+++ source/databricks/calculation_engine/package/calculation/energy/energy_calculation.py
@@ -17,7 +17,7 @@ import package.calculation.energy.aggregators.exchange_aggregators as exchange_a
 import package.calculation.energy.aggregators.grid_loss_aggregators as grid_loss_aggr
 import package.calculation.energy.aggregators.grouping_aggregators as grouping_aggr
 import package.calculation.energy.aggregators.metering_point_time_series_aggregators as mp_aggr
-import package.calculation.output.calculation_results.energy_storage_model_factory as factory
+import package.calculation.output.results.energy_storage_model_factory as factory
 from package.calculation.calculation_results import EnergyResultsContainer
 from package.calculation.calculator_args import CalculatorArgs
 from package.calculation.energy.data_structures.energy_results import EnergyResults
