+++ source/databricks/calculation_engine/package/calculation/output/results/schemas/energy_results_schema.py
