+++ source/databricks/calculation_engine/package/calculation/output/results/schemas/total_monthly_amounts_schema.py
