+++ source/databricks/calculation_engine/package/calculation/output/results/total_monthly_amounts_storage_model_factory.py
@@ -16,7 +16,7 @@ from pyspark.sql import DataFrame
 from pyspark.sql.functions import col
 
 from package.calculation.calculator_args import CalculatorArgs
-from package.calculation.output.calculation_results.add_meta_data import add_metadata
+from package.calculation.output.results.add_meta_data import add_metadata
 from package.calculation.wholesale.data_structures import TotalMonthlyAmount
 from package.constants import Colname, TotalMonthlyAmountsColumnNames
