+++ source/databricks/calculation_engine/tests/calculation/output/basis_data/basis_data_test_factory.py
@@ -19,7 +19,7 @@ from pyspark.sql import Row, SparkSession, DataFrame
 import package.calculation.output.basis_data.basis_data_factory as basis_data_factory
 import package.codelists as e
 import tests.calculation.charges_factory as charges_factory
-from calculation.output.calculations_storage_model_test_factory import (
+from calculation.output.results.calculations_storage_model_test_factory import (
     create_calculations,
 )
 from package.calculation.output.basis_data.schemas.charge_link_periods_schema import (
