+++ source/databricks/calculation_engine/tests/calculation/output/results/test_wholesale_storage_model_factory.py
@@ -20,7 +20,7 @@ import pytest
 from pyspark.sql import SparkSession, DataFrame
 
 from package.calculation.calculator_args import CalculatorArgs
-from package.calculation.output.calculation_results import (
+from package.calculation.output.results import (
     wholesale_storage_model_factory as sut,
 )
 from package.calculation.wholesale.data_structures.wholesale_results import (
