+++ source/databricks/calculation_engine/tests/datamigration/subsystem/test_energy_results_migrations.py
@@ -21,7 +21,7 @@ from pyspark.sql import SparkSession, DataFrame
 from pyspark.sql.functions import lit, col
 
 from helpers.data_frame_utils import set_column
-from package.calculation.output.calculation_results.schemas import (
+from package.calculation.output.results.schemas import (
     hive_energy_results_schema,
 )
 from package.codelists import (
