+++ source/databricks/calculation_engine/tests/datamigration/subsystem/test_monthly_amounts_per_charge_migrations.py
@@ -19,7 +19,7 @@ import pytest
 from pyspark.sql import SparkSession, DataFrame
 from pyspark.sql.functions import col, lit
 
-from package.calculation.output.calculation_results.schemas import (
+from package.calculation.output.results.schemas import (
     monthly_amounts_schema,
 )
 from package.codelists import CalculationType, ChargeType
